pyproject.toml
[project]
name = "cycling-trip-planner"
version = "0.1.0"
description = "AI agent for planning multi-day cycling trips"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "anthropic>=0.42.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.38.0",
    "requests>=2.31.0",
    "folium>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
